Working on cleaning duplicated code
