+++ main.js
@@ -1,3 +1,6 @@
+let {json_data} = require("./public/scripts/main_setup");
+const {chapters} = require("./public/scripts/main_setup");
+
 let ejs = require('ejs');
 const path = require('path')
 
@@ -23,6 +26,8 @@ app.on('ready', function() {
   mainWindow.loadURL(`http://localhost:${port}/`);
   mainWindow.focus();
 });
+
+
 /* Initial Logic to get the header from the id.
 TODO: Discuss possibilities of alternate logic that do not need inference, 
   but can simply be stored in the JSON object itself.
@@ -49,26 +54,6 @@ function render_scene(req,json_data,id) {
     data.text = ejs.render(data.text,req.cookies);
     return ejs.renderFile(path.join(__dirname,"templates/main.ejs"),data)
 }
-
-// Temporary array to keep track of stats variables
-const stats_variables = [
-  "v_available_points",
-  "v_strength",
-  "v_toughness",
-  "v_agility", // Speed
-  "v_reflexes",
-  "v_hearing",
-  "v_perception", // Observation
-  "v_ancient_languages",
-  "v_combat_technique",
-  "v_premonition",
-  "v_bluff",
-  "v_magical_sense",
-  "v_aura_hardening",
-  "v_magical_power", // Currently, not utilized
-  "v_magical_knowledge", // Currently, not utilized
-  "v_max_stat"
-]
 
 function render_stats(req) {
     return ejs.renderFile(path.join(__dirname,"templates/stats.ejs"),req.cookies)
@@ -100,12 +85,9 @@ function render_saves(req) {
     return ejs.renderFile(path.join(__dirname,"templates/saves.ejs"),data)
 }
 
-let json_data = {}
-const chapters = (
-    ["ch1","ch2","ch3","ch4","ch5","ch6","ch7","ch8","ch9","ch10","ch11a","ch11b",
-    'b2ch1', 'b2ch2', 'b2ch3', 'b2ch4a', 'b2ch4b', 'b2ch5a', 'b2ch5b', 'b2ch6', 'b2ch7','b2ch8', 'b2ch9a', 'b2ch9b', 'b2ch10a', 'b2ch10b', 'b2ch11a', 'b2ch11b', 'b2ch11c',
-    'b3ch1', 'b3ch2a', 'b3ch2b', 'b3ch2c', 'b3ch3a', 'b3ch3b', 'b3ch4a', 'b3ch4b', 'b3ch5a', 'b3ch5b', 'b3ch6a', 'b3ch6b', 'b3ch6c', 'b3ch7a', 'b3ch8a', 'b3ch8b', 'b3ch9a', 'b3ch9b', 'b3ch9c', 'b3ch10a', 'b3ch10b', 'b3ch10c', 'b3ch11a', 'b3ch12a', 'b3ch12b']
-)
+
+/// ---
+
 for (chapter of chapters) {
     json_data = Object.assign(json_data,require(path.join(__dirname,`chapters/${chapter}.json`)))
 }
@@ -125,6 +107,7 @@ expressApp.use(cookieParser());
 // Serve static files
 expressApp.use(express.static(path.join(__dirname, 'public')));
 
+// TODO: Move those inside `main_setup.js`?
 expressApp.get('/', (req, res) => {
     const id = req.cookies.v_current_scene ? req.cookies.v_current_scene : "Ch1-Intro1"
     render_full(req,(r) => render_scene(r,json_data[id],get_header_from_id(id)),get_header_from_id(id)).then((rendered) => res.send(rendered))
