+++ main_node.js
@@ -1,3 +1,6 @@
+let {json_data} = require("./public/scripts/main_setup");
+const {chapters} = require("./public/scripts/main_setup");
+
 let ejs = require('ejs');
 
 /* Initial Logic to get the header from the id.
@@ -27,26 +30,6 @@ function render_scene(req,json_data,id) {
     return ejs.renderFile("templates/main.ejs",data)
 }
 
-// Temporary array to keep track of stats variables
-const stats_variables = [
-  "v_available_points",
-  "v_strength",
-  "v_toughness",
-  "v_agility", // Speed
-  "v_reflexes",
-  "v_hearing",
-  "v_perception", // Observation
-  "v_ancient_languages",
-  "v_combat_technique",
-  "v_premonition",
-  "v_bluff",
-  "v_magical_sense",
-  "v_aura_hardening",
-  "v_magical_power", // Currently, not utilized
-  "v_magical_knowledge", // Currently, not utilized
-  "v_max_stat"
-]
-
 function render_stats(req) {
     return ejs.renderFile("templates/stats.ejs",req.cookies)
 }
@@ -67,9 +50,8 @@ function render_achievements_menu_chapter(req,achievements_data) {
     return ejs.renderFile("templates/achievements_menu_chapter.ejs",Object.assign({},{"achievements":achievements_data},req.cookies))
 }
 
-
 function render_saves(req) {
-    let saveData = {} 
+    let saveData = {}
     Object.entries(req.body).forEach(function(entry){
         saveData[entry[0]] = {"date": entry[1].date}
     })
@@ -77,12 +59,8 @@ function render_saves(req) {
     return ejs.renderFile("templates/saves.ejs",data)
 }
 
-let json_data = {}
-const chapters = (
-    ["ch1","ch2","ch3","ch4","ch5","ch6","ch7","ch8","ch9","ch10","ch11a","ch11b",
-    'b2ch1', 'b2ch2', 'b2ch3', 'b2ch4a', 'b2ch4b', 'b2ch5a', 'b2ch5b', 'b2ch6', 'b2ch7','b2ch8', 'b2ch9a', 'b2ch9b', 'b2ch10a', 'b2ch10b', 'b2ch11a', 'b2ch11b', 'b2ch11c',
-    'b3ch1', 'b3ch2a', 'b3ch2b', 'b3ch2c', 'b3ch3a', 'b3ch3b', 'b3ch4a', 'b3ch4b', 'b3ch5a', 'b3ch5b', 'b3ch6a', 'b3ch6b', 'b3ch6c', 'b3ch7a', 'b3ch8a', 'b3ch8b', 'b3ch9a', 'b3ch9b', 'b3ch9c', 'b3ch10a', 'b3ch10b', 'b3ch10c', 'b3ch11a', 'b3ch12a', 'b3ch12b']
-)
+/// ---
+
 for (chapter of chapters) {
     json_data = Object.assign(json_data,require(`./chapters/${chapter}.json`))
 }
@@ -103,6 +81,7 @@ const port = 3000
 const path = require('path')
 app.use(express.static(path.join(__dirname, 'public')));
 
+// TODO: Move those inside `main_setup.js`?
 app.get('/', (req, res) => {
     const id = req.cookies.v_current_scene ? req.cookies.v_current_scene : "Ch1-Intro1"
     render_full(req,(r) => render_scene(r,json_data[id],get_header_from_id(id)),get_header_from_id(id)).then((rendered) => res.send(rendered))
@@ -117,12 +96,12 @@ app.get('/scene/:id', (req, res) => {
     render_full(req,callback,get_header_from_id(req.params.id)).then((rendered) => res.send(rendered));
 })
 
-app.get('/stats', (req, res) => {  
+app.get('/stats', (req, res) => {
     render_full(req,render_stats,"Stats").then((rendered) => res.send(rendered));
 })
 
 
-app.get('/achievements', (req, res) => {  
+app.get('/achievements', (req, res) => {
     render_full(req,render_achievements_menu,"Achievements").then((rendered) => res.send(rendered));
 })
 
@@ -137,7 +116,7 @@ app.get('/achievements/book/(:idBook)/chapter/(:idChapter)', (req, res) => {
     render_full(req,callback,"Achievements").then((rendered) => res.send(rendered))
 })
 
-app.all('/saves', bodyParser.json(), (req, res) => {  
+app.all('/saves', bodyParser.json(), (req, res) => {
     render_full(req,render_saves,"Save files").then((rendered) => res.send(rendered));
 })
